% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_module_address_entry.R
\name{addressEntryServer}
\alias{addressEntryServer}
\title{Address Entry Module Server}
\usage{
addressEntryServer(id, config, db)
}
\arguments{
\item{id}{The module ID}

\item{config}{The path to a config file ingested by \code{{config}}}

\item{db}{The database connection pool}
}
\value{
The server for the address entry module
}
\description{
Server for the address entry module
}
