% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_module_address_entry.R
\name{addressEntryUI}
\alias{addressEntryUI}
\title{Address Entry Module UI}
\usage{
addressEntryUI(id)
}
\arguments{
\item{id}{The module ID}
}
\value{
The UI for the address entry module
}
\description{
UI for the address entry module
}
