% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mturk_assignment.R
\name{assignment_record_exists}
\alias{assignment_record_exists}
\title{Assignment Record Exists}
\usage{
assignment_record_exists(db, assignment)
}
\arguments{
\item{db}{A database connection pool}

\item{assignment}{The Assignment ID}
}
\description{
Assignment Record Exists
}
