% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{batch_case}
\alias{batch_case}
\title{Batch a Case}
\usage{
batch_case(db, hit_type)
}
\arguments{
\item{db}{A database connection pool}

\item{hit_type}{The HIT Type ID to use for this batch}
}
\description{
Batch a Case
}
