% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{calculate_pay}
\alias{calculate_pay}
\title{Calculate Pay}
\usage{
calculate_pay(
  db,
  users = "all",
  start = get_pay_period(lubridate::today())$start,
  end = get_pay_period(lubridate::today())$end,
  priority_rate = 0.35,
  backlog_rate = 0.25
)
}
\arguments{
\item{db}{A database connection pool created with \code{pool::dbPool}}

\item{users}{A vector of users to plot}

\item{start}{The start date}

\item{end}{The end date}
}
\value{
A tibble with the number of addresses entered by type and the corresponding pay
}
\description{
Calculates the pay for a given user
}
\examples{
\dontrun{
calculate_pay(db, c("test", "test2"), lubridate::ymd("2022-12-12"), lubridate::today())
}

}
