% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_auth.R
\name{change_password}
\alias{change_password}
\title{Change Password}
\usage{
change_password(db, username, old_password, new_password)
}
\arguments{
\item{db}{A database connection pool created with \code{pool::dbPool}}

\item{username}{The username}

\item{old_password}{The old password}

\item{new_password}{The new password}
}
\value{
Returns invisibly if successful
}
\description{
Changes a user's password
}
\examples{
\dontrun{
change_password(db, "test", "test", "test2")
}

}
