% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgrid.R
\name{check_line_args}
\alias{check_line_args}
\title{Check Line Arguments}
\usage{
check_line_args(line1, line2)
}
\arguments{
\item{line1}{Line 1 of the address}

\item{line2}{Line 2 of the address}
}
\description{
Check Line Arguments
}
