% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgrid.R
\name{check_street_args}
\alias{check_street_args}
\title{Check Street Arguments}
\usage{
check_street_args(
  street_number,
  street_direction,
  street_name,
  street_type,
  unit
)
}
\arguments{
\item{street_number}{Street number}

\item{street_direction}{Street direction}

\item{street_name}{Street name}

\item{street_type}{Street type}

\item{unit}{Unit}
}
\description{
Check Street Arguments
}
