% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{clean_queue}
\alias{clean_queue}
\title{Clean Queue}
\usage{
clean_queue(db)
}
\arguments{
\item{db}{A database connection}
}
\description{
Removes completed cases from the queue and handles abandoned cases
}
