% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_module_current_documents.R
\name{currentDocumentsServer}
\alias{currentDocumentsServer}
\title{Current Documents Module Server}
\usage{
currentDocumentsServer(id, current_case, db)
}
\arguments{
\item{id}{The id of the module}

\item{current_case}{The reactive value containing the current case}

\item{db}{The database connection}
}
\value{
The server function for the current documents module
}
\description{
This function creates the server for the current documents module.
}
