% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_module_current_documents.R
\name{currentDocumentsUI}
\alias{currentDocumentsUI}
\title{Current Documents Module UI}
\usage{
currentDocumentsUI(id)
}
\arguments{
\item{id}{The id of the module}
}
\value{
A tagList of the UI
}
\description{
This function creates the UI for the current documents module.
}
