% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_server.R
\name{dashboard_server}
\alias{dashboard_server}
\title{Eviction Addresses Dashboard Server}
\usage{
dashboard_server(connection_args)
}
\arguments{
\item{input}{The server input}

\item{output}{The server output}

\item{session}{The server session}
}
\value{
A Shiny server function
}
\description{
The server function passed to \code{shinyApp}
}
