% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_ui.R
\name{dashboard_ui}
\alias{dashboard_ui}
\title{Eviction Addresses Dashboard UI}
\usage{
dashboard_ui(cookie_expiry)
}
\arguments{
\item{cookie_expiry}{The expiration length for the session cookie}
}
\value{
A shiny ui function
}
\description{
The ui function passed to \code{shinyApp}
}
