% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{download_oscn_document}
\alias{download_oscn_document}
\title{Download OSCN Document}
\usage{
download_oscn_document(id, link)
}
\arguments{
\item{id}{The id of the document minute}

\item{link}{The link to the document}
}
\value{
A pdf document as a raw vector
}
\description{
Download a document from OSCN
}
