% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{email_pay_report}
\alias{email_pay_report}
\title{Email Pay Report}
\usage{
email_pay_report(db, users, start, end, recipient_email)
}
\arguments{
\item{db}{A database connection pool created with \code{pool::dbPool}}

\item{users}{A vector of users to plot}

\item{start}{The start date}

\item{end}{The end date}

\item{email}{The email address to send the report to}
}
\value{
Nothing
}
\description{
Emails a pay report for the given users using the package \code{{blastula}}
}
