% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{email_project_report}
\alias{email_project_report}
\title{Email Project Report}
\usage{
email_project_report(
  db,
  start = get_pay_period(lubridate::today())$start,
  end = get_pay_period(lubridate::today())$end,
  recipient_email,
  ...,
  .silent = FALSE
)
}
\arguments{
\item{db}{A database connection pool created with \code{pool::dbPool}}

\item{start}{The start date}

\item{end}{The end date}

\item{recipient_email}{The email address to send the report to}

\item{...}{Additional arguments placeholder}

\item{.silent}{A boolean indicating whether to print the email}
}
\value{
Nothing
}
\description{
Emails a project report to a given user
}
