% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_module_entry_detail.R
\name{entryDetailServer}
\alias{entryDetailServer}
\title{Entry Detail Module Server}
\usage{
entryDetailServer(id, current_case, total_cases)
}
\arguments{
\item{id}{The module ID}

\item{current_case}{The reactive value containing the current case}

\item{total_cases}{The reactive value containing the total number of cases}
}
\value{
The server function for the entry detail module
}
\description{
Server function for the entry detail module
}
