% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_module_entry_detail.R
\name{entryDetailUI}
\alias{entryDetailUI}
\title{Entry Detail Module UI}
\usage{
entryDetailUI(id)
}
\arguments{
\item{id}{The module ID}
}
\value{
The UI function for the entry detail module
}
\description{
UI function for the entry detail module
}
