% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expect_no_error}
\alias{expect_no_error}
\title{Expect No Error}
\usage{
expect_no_error(object)
}
\arguments{
\item{object}{An object passed to \code{testthat::expect_error()}}
}
\description{
Expect No Error
}
