% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mturk.R
\name{finalize_hit}
\alias{finalize_hit}
\title{Finalize HIT}
\usage{
finalize_hit(hit = NULL)
}
\arguments{
\item{hit}{The HIT id to finalize}
}
\value{
Nothing
}
\description{
Finalize HIT
}
\examples{

finalize_hit(hit = "<insert hit id>")

}
