% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgrid.R
\name{format_postgrid_request}
\alias{format_postgrid_request}
\title{Format PostGrid Request}
\usage{
format_postgrid_request(
  line1 = NA_character_,
  line2 = NA_character_,
  city = NA_character_,
  state = "ok",
  zip = NA_character_,
  country = "us",
  street_number = NA_character_,
  street_direction = NA_character_,
  street_name = NA_character_,
  street_type = NA_character_,
  unit = NA_character_
)
}
\arguments{
\item{line1}{The first line of an address}

\item{line2}{The second line of an address (e.g. unit/suite/apartment)}

\item{city}{The city of the address}

\item{state}{The state of the address}

\item{zip}{The five-digit zip code of the address}

\item{country}{The two letter country code}

\item{street_number}{The building number of the address's street}

\item{street_direction}{The street direction of the address}

\item{street_name}{The name of the address's street (including post-direction, e.g. 68th E)}

\item{street_type}{The type of the address's street (e.g. street/avenue/place)}

\item{unit}{The unit type and value of an address (e.g. APT 13, UNIT C)}
}
\value{
A list with fields suitable for a PostGrid request
}
\description{
Format PostGrid Request
}
