% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mturk_assignment.R
\name{get_assignment_answer}
\alias{get_assignment_answer}
\title{Get Assignment Answer}
\usage{
get_assignment_answer(assignment)
}
\arguments{
\item{assignment}{The Assignment id. A string (character vector length one)}
}
\value{
The parsed assignment as a ??
}
\description{
Get Assignment Answer
}
