% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mturk.R
\name{get_assignment_status}
\alias{get_assignment_status}
\title{Get Assignment Status}
\usage{
get_assignment_status(assignment)
}
\arguments{
\item{assignment}{The assignment id. A string (character vector length one).}
}
\value{
The assignment status. A string (character vector length one). See \code{valid_assignment_statuses} for possible values.
}
\description{
Get Assignment Status
}
