% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{get_case_from_queue}
\alias{get_case_from_queue}
\title{Get Case from Queue}
\usage{
get_case_from_queue(db)
}
\arguments{
\item{db}{A database connection pool created with \code{pool::dbPool}}
}
\value{
A case id
}
\description{
Get Case from Queue
}
