% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_documents_by_case.R
\name{get_documents_by_case}
\alias{get_documents_by_case}
\title{Get Documents by Case}
\usage{
get_documents_by_case(id)
}
\arguments{
\item{id}{The id of the case for which to return documents}
}
\value{
A data.frame of documents
}
\description{
Get Documents by Case
}
