% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mturk_assignment.R
\name{get_hit_assignments}
\alias{get_hit_assignments}
\title{Get HIT Assignments}
\usage{
get_hit_assignments(db, hit)
}
\arguments{
\item{db}{A database connection pool}

\item{hit}{The HIT id. A string (character vector length one)}
}
\value{
A character vector of Assignment ids
}
\description{
Get HIT Assignments
}
