% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mturk_hit.R
\name{get_hit_status}
\alias{get_hit_status}
\title{Get HIT Status}
\usage{
get_hit_status(hit)
}
\arguments{
\item{hit}{The hit id. A string (character vector of length one)}
}
\value{
The HIT status. A string (character vector of length one). See \code{valid_hit_statuses} for possible values.
}
\description{
Get HIT Status
}
