% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sessions_from_db.R
\name{get_sessions_from_db}
\alias{get_sessions_from_db}
\title{Get Sessions from DB}
\usage{
get_sessions_from_db(conn = db, expiry = cookie_expiry)
}
\arguments{
\item{conn}{The database connection}

\item{expiry}{The cookie expiration}
}
\value{
A tibble of session info
}
\description{
Gets a tibble of sessions from the database for use by {shinyauth}
}
