% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_pings.R
\name{handle_dbping}
\alias{handle_dbping}
\title{Handle DB Ping}
\usage{
handle_dbping(db)
}
\arguments{
\item{db}{A database connection pool created with \code{pool::dbPool}}
}
\value{
Empty
}
\description{
A plumber handler that pings the database
}
