% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_hydrate.R
\name{handle_hydrate}
\alias{handle_hydrate}
\title{Handle Hydrate}
\usage{
handle_hydrate(db)
}
\value{
A 200, if successful
}
\description{
A plumber handler for the /hydrate endpoint
}
\details{
This endpoint searches eviction document minutes for links. It downloads the
documents, stores them in a Google Cloud bucket, then creates an internal
link to the document.
}
