% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_mturk_batch.R
\name{handle_mturk_batch}
\alias{handle_mturk_batch}
\title{Handle MTurk Batch}
\usage{
handle_mturk_batch(db, max_batch_size)
}
\arguments{
\item{db}{A database connection}
}
\value{
A handler function for \code{plumber::pr_handle}
}
\description{
Handle MTurk Batch
}
