% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_mturk_review.R
\name{handle_mturk_review}
\alias{handle_mturk_review}
\title{Handle MTurk Review}
\usage{
handle_mturk_review()
}
\value{
A handler function for plumber endpoint /mturk/review
}
\description{
Handle MTurk Review
}
