% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_pings.R
\name{handle_ping}
\alias{handle_ping}
\title{Handle Ping}
\usage{
handle_ping()
}
\value{
A plumber handler that returns a 200 status code and a message that the api is connected
}
\description{
A plumber handler for a simple ping of the api
}
\details{
This endpoint returns a simple "pong" message
}
