% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_refresh.R
\name{handle_refresh}
\alias{handle_refresh}
\title{Handle API Refresh}
\usage{
handle_refresh(config)
}
\value{
A 200 if successful
}
\description{
Plumber handler for endpoint /refresh
}
\details{
This endpoint refreshes materialized views and inserts new cases and
documents into the eviction_addresses schema. It then updates the work queue
based on what it finds.
}
