% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_refresh.R
\name{handle_refresh_queue}
\alias{handle_refresh_queue}
\title{Handle API Refresh Queue}
\usage{
handle_refresh_queue(db)
}
\arguments{
\item{db}{A database connection pool}
}
\description{
Plumber handler for endpoint \verb{/refresh/queue}
}
\details{
This endpoint refreshes the work queue based on what it finds in the
eviction_addresses schema.
}
