% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has_latlon_match}
\alias{has_latlon_match}
\title{Has Lat/Lon Match}
\usage{
has_latlon_match(.data, tolerance = 2e-04)
}
\arguments{
\item{.data}{A tibble with columns \code{lat} and \code{lon}}

\item{tolerance}{A numeric >= 0}
}
\value{
A boolean
}
\description{
Has Lat/Lon Match
}
