% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has_names}
\alias{has_names}
\title{Has Names}
\usage{
has_names(x, names)
}
\arguments{
\item{x}{The object to check for names}

\item{names}{A character vector of names to check \code{x} for}
}
\value{
A boolean (logical vector of length 1)
}
\description{
Has Names
}
