% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boto3.R
\name{install_boto3}
\alias{install_boto3}
\title{Install Boto3}
\usage{
install_boto3(method = "auto", conda = "auto")
}
\arguments{
\item{method}{A string (character vector length one)}

\item{conda}{A string (character vector length one)}
}
\description{
Install Boto3
}
