% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_module_address_entry.R
\name{isolate_address_entered}
\alias{isolate_address_entered}
\title{Isolate Address Entered}
\usage{
isolate_address_entered(input)
}
\arguments{
\item{input}{The input object}
}
\description{
This function isolates the address entered by the user.
}
