% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mturk.R
\name{mturk_auth}
\alias{mturk_auth}
\title{MTurk Authorization}
\usage{
mturk_auth(config = NULL)
}
\arguments{
\item{config}{A file path to a config.yml}
}
\value{
A logical/boolean representing whether we successfully connected to MTurk
}
\description{
MTurk Authorization
}
\examples{

\dontrun{
try(mturk_auth())
try(mturk_auth(config = "config.yml"))
}

}
