% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mturk_assignment.R
\name{new_assignment_record}
\alias{new_assignment_record}
\title{New Assignment Record}
\usage{
new_assignment_record(db, hit, assignment, worker, status)
}
\arguments{
\item{db}{A database connection pool}

\item{hit}{A HIT ID}

\item{assignment}{The Assignment ID}

\item{worker}{The Worker ID associated with the Assignment}

\item{status}{The status of the Assignment to record}
}
\description{
New Assignment Record
}
