% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mturk.R
\name{new_case_from_queue}
\alias{new_case_from_queue}
\title{New Case from Queue}
\usage{
new_case_from_queue()
}
\value{
A case id from the queue
}
\description{
New Case from Queue
}
\examples{

new_case_from_queue()

}
