% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_db_connection.R
\name{new_db_connection}
\alias{new_db_connection}
\title{New DB Connection}
\usage{
new_db_connection(config = "config.yml")
}
\arguments{
\item{connection_args}{A list of arguments to be used in the {pool} configuration. This should be created by the {config} package using \code{config::get()} and a \code{config.yml} file}
}
\value{
A database connection pool created by the {pool} package
}
\description{
Creates a new connection to the database
}
