% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mturk_hit.R
\name{new_hit_record}
\alias{new_hit_record}
\title{New HIT Record}
\usage{
new_hit_record(db, hit, case)
}
\arguments{
\item{db}{A database connection pool}

\item{hit}{The HIT ID to record}

\item{case}{The case id associated with the HIT}
}
\description{
New HIT Record
}
