% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mturk.R
\name{new_hit_type}
\alias{new_hit_type}
\title{New HIT Type}
\usage{
new_hit_type(
  title = "eviction-address-transcription",
  description = "Find and transcribe the DEFENDENT'S address from a court document pdf",
  reward = "0.15",
  duration = pyMTurkR::seconds(minutes = 10),
  keywords = "address, text, transcribe, entry, data",
  auto.approval.delay = pyMTurkR::seconds(days = 3),
  ...
)
}
\arguments{
\item{title}{The title for the HIT Type}

\item{description}{A description to be shown to workers}

\item{reward}{A string representing amount to be paid for successful completion in USD}

\item{duration}{The time in seconds that a worker has to complete the assignment}

\item{keywords}{A string with comma separated terms to be used to search for HITs of this type}

\item{auto.approval.delay}{The time in seconds after which a completed assignment is auto-approved}

\item{...}{Additional arguments passed to \code{pyMTurkR::CreateHITType()}}
}
\value{
The HIT Type id
}
\description{
New HIT Type
}
\examples{

\dontrun{
new_hit_type()
new_hit_type(reward = "0.20")
new_hit_type(duration = pyMTurkR::seconds(minutes = 20))
}

}
