% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{new_mturk_batch}
\alias{new_mturk_batch}
\title{New MTurk Batch}
\usage{
new_mturk_batch(db, max_batch_size, hit_type)
}
\arguments{
\item{db}{A database connection pool}

\item{max_batch_size}{The maximum number of HITs to create at once}

\item{hit_type}{The HIT Type to use for created HITs}
}
\value{
Empty
}
\description{
New MTurk Batch
}
