% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_auth.R
\name{new_user}
\alias{new_user}
\title{New User}
\usage{
new_user(db, username, password, role, name)
}
\arguments{
\item{db}{A database connection pool created with \code{pool::dbPool}}

\item{username}{The username}

\item{password}{The password}

\item{role}{The role of the user; either "admin" or "standard"}

\item{name}{The name of the user}
}
\value{
Returns invisibly if successful
}
\description{
Adds a new user to the database
}
\examples{
\dontrun{
new_user(db, "test", "test", "admin", "Test User")
}

}
