% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_module_address_entry.R
\name{observe_address_submission}
\alias{observe_address_submission}
\title{Observe Address Submission}
\usage{
observe_address_submission(
  input,
  db,
  current_case,
  current_user,
  address_entered,
  address_validated
)
}
\arguments{
\item{input}{The input object}

\item{db}{The database connection}

\item{current_case}{The current case}

\item{current_user}{The current user}

\item{address_entered}{The address entered by the user}

\item{address_validated}{The address validated by the API}
}
\value{
A Shiny observer object
}
\description{
This function observes the address submission button.
}
