% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_module_address_entry.R
\name{observe_address_validation}
\alias{observe_address_validation}
\title{Observe Address Validation}
\usage{
observe_address_validation(
  input,
  current_case,
  jwt,
  api_url,
  address_entered,
  address_validated
)
}
\arguments{
\item{input}{The input object from the Shiny app}

\item{current_case}{The reactive value for the current case}

\item{jwt}{The JWT token for the API}

\item{api_url}{The URL for the API}

\item{address_entered}{The reactive values for the address entered}

\item{address_validated}{The reactive values for the address validated}
}
\value{
A Shiny observeEvent object
}
\description{
This function observes the address validation button and
validates the address.
}
