% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mturk.R
\name{parse_assignment_answer}
\alias{parse_assignment_answer}
\title{Parse HIT Assignment Answer}
\usage{
parse_assignment_answer(answer)
}
\arguments{
\item{answer}{A data.frame containing the answer for one HIT assignment}
}
\value{
A parsed address ready for validation
}
\description{
Parse HIT Assignment Answer
}
