% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mturk.R
\name{parse_assignment_answers}
\alias{parse_assignment_answers}
\title{Parse HIT Assignment Answers}
\usage{
parse_assignment_answers(answers)
}
\arguments{
\item{answers}{A data.frame containing the answers for one HIT assignment}
}
\value{
A parsed address ready for validation
}
\description{
Parse HIT Assignment Answers
}
