% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgrid.R
\name{parse_postgrid_response}
\alias{parse_postgrid_response}
\title{Parse PostGrid Response}
\usage{
parse_postgrid_response(res)
}
\arguments{
\item{res}{A response from the Postgrid API}
}
\value{
A list with values specified according to the PostGrid documentation
}
\description{
Parse PostGrid Response
}
