% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{plot_address_coverage}
\alias{plot_address_coverage}
\title{Plot Address Coverage}
\usage{
plot_address_coverage(db, ..., .silent = FALSE)
}
\arguments{
\item{db}{A database connection pool created with \code{pool::dbPool}}

\item{...}{Additional arguments placeholder}

\item{.silent}{A boolean indicating whether to print the plot}
}
\value{
A tibble with the number of addresses entered by month
}
\description{
Calculates the number of addresses entered
}
