% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_auth.R
\name{plot_address_entries}
\alias{plot_address_entries}
\title{Plot Address Entries}
\usage{
plot_address_entries(
  db,
  users,
  start = lubridate::ymd("2022-12-12"),
  end = lubridate::today(),
  type
)
}
\arguments{
\item{db}{A database connection pool created with \code{pool::dbPool}}

\item{users}{A vector of users to plot}

\item{start}{The start date}

\item{end}{The end date}

\item{type}{The type of address entry to plot; either "priority" or "backlog"}
}
\value{
A ggplot object
}
\description{
Plots the number of address entries per day
}
\examples{
\dontrun{
plot_address_entries(db, c("test", "test2"), lubridate::ymd("2022-12-12"), lubridate::today(), "priority")
}

}
