% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{plot_cases}
\alias{plot_cases}
\title{Plot Cases}
\usage{
plot_cases(db, ..., .silent = FALSE)
}
\arguments{
\item{db}{A database connection pool created with \code{pool::dbPool}}

\item{...}{Additional arguments placeholder}

\item{.silent}{A boolean indicating whether to print the plot}
}
\value{
A ggplot object
}
\description{
Calculates the number of cases filed per month
}
