% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{plot_entries}
\alias{plot_entries}
\title{Plot Entries}
\usage{
plot_entries(db, ..., .silent = FALSE)
}
\arguments{
\item{db}{A database connection pool created with \code{pool::dbPool}}

\item{...}{Additional arguments placeholder}

\item{.silent}{A boolean indicating whether to print the plot}
}
\value{
A ggplot object
}
\description{
Calculates the number of addresses entered
}
