% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case.R
\name{refresh_cases}
\alias{refresh_cases}
\title{Refresh Cases}
\usage{
refresh_cases(db)
}
\arguments{
\item{db}{A database connection}
}
\description{
Refreshes the cases table with new cases from the recent_tulsa_evictions view
}
