% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{refresh_documents}
\alias{refresh_documents}
\title{Refresh Documents}
\usage{
refresh_documents(db, n = NULL)
}
\arguments{
\item{db}{A database connection}

\item{n}{The number of documents to refresh}
}
\description{
Refresh Documents
}
