% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mturk_layout.R
\name{render_document_links}
\alias{render_document_links}
\title{Render Document Links}
\usage{
render_document_links(links)
}
\arguments{
\item{links}{A character vector of document links}
}
\value{
An HTML string
}
\description{
Render Document Links
}
\examples{

links <- c(
  "https://google.com",
  "https://google.com"
)
render_document_links(links)

}
