% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mturk_layout.R
\name{render_hit_layout_for_case}
\alias{render_hit_layout_for_case}
\title{Render HIT Layout for Case}
\usage{
render_hit_layout_for_case(db, case, layout = NULL)
}
\arguments{
\item{db}{A database pool created with \code{pool::dbPool}}

\item{case}{A case id used to render the layout}

\item{layout}{A file path to an XML layout}
}
\value{
A character string containing an XML layout
}
\description{
Render HIT Layout for Case
}
