% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{render_pay_report}
\alias{render_pay_report}
\title{Render Pay Report}
\usage{
render_pay_report(db, users, start, end)
}
\arguments{
\item{db}{A database connection pool created with \code{pool::dbPool}}

\item{users}{A vector of users to plot}

\item{start}{The start date}

\item{end}{The end date}
}
\value{
A tibble with the number of addresses entered by type and the corresponding pay
}
\description{
Renders a pay report for a given user
}
