% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{render_project_report}
\alias{render_project_report}
\title{Render Project Report}
\usage{
render_project_report(
  db,
  start = get_pay_period(lubridate::today())$start,
  end = get_pay_period(lubridate::today())$end,
  ...,
  .silent = FALSE
)
}
\arguments{
\item{db}{A database connection pool created with \code{pool::dbPool}}

\item{start}{The start date}

\item{end}{The end date}
}
\value{
A rendered project report
}
\description{
Renders a project report for a given user
}
