% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mturk_hit.R
\name{review_hit}
\alias{review_hit}
\title{Review HIT}
\usage{
review_hit(db, config, hit)
}
\arguments{
\item{db}{A database connection pool}

\item{config}{A configuration file}

\item{hit}{The HIT id. A string (character vector length one)}
}
\description{
Review HIT
}
