% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mturk_assignment.R
\name{review_hit_assignments}
\alias{review_hit_assignments}
\title{Review HIT Assignments}
\usage{
review_hit_assignments(hit)
}
\arguments{
\item{hit}{The HIT id}
}
\description{
Review HIT Assignments
}
