% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_job.R
\name{run_job}
\alias{run_job}
\title{Run Job}
\usage{
run_job(config)
}
\arguments{
\item{config}{The path to a configuration file ingested by \code{{config}}}
}
\value{
Nothing
}
\description{
Runs a refresh job
}
