% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgrid.R
\name{send_postgrid_request}
\alias{send_postgrid_request}
\title{Send PostGrid Request}
\usage{
send_postgrid_request(config = NULL, address = list(), geocode = T)
}
\arguments{
\item{config}{The path of a \code{config.yml} file with section to be parsed by \code{config::get(value = "postgrid")}}

\item{address}{A list with elements \code{line1}, \code{line2}, \code{city}, \code{provinceOrState}, and \code{country}}

\item{geocode}{A flag (logical vector of length one) indicating whether to geocode the address. Uses another Postgrid unit. Defaults to \code{TRUE}}
}
\value{
A PostGrid response
}
\description{
Send PostGrid Request
}
