% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve_dashboard.R
\name{serve_dashboard}
\alias{serve_dashboard}
\title{Serve Dashboard}
\usage{
serve_dashboard(port = 3838)
}
\arguments{
\item{port}{The port to serve the dashboard}
}
\description{
Serves the eviction address entry dashboard
}
