% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mturk.R
\name{set_hit_type}
\alias{set_hit_type}
\title{Set HIT Type}
\usage{
set_hit_type(
  title = "eviction-address-transcription",
  description = "Find and transcribe the DEFENDENT'S address from a court document pdf",
  reward = "0.15",
  duration = pyMTurkR::seconds(minutes = 10),
  keywords = "address, text, transcribe, entry, data",
  auto.approval.delay = pyMTurkR::seconds(days = 3),
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Set HIT Type
}
