% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{store_public_document_link}
\alias{store_public_document_link}
\title{Store Public Document Link}
\usage{
store_public_document_link(db, id)
}
\arguments{
\item{db}{A database connection pool created with \code{pool::dbPool}}

\item{id}{The id of the document}
}
\description{
Store the GCS link to a document in the database
}
