% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_module_address_entry.R
\name{stringify_address_entered}
\alias{stringify_address_entered}
\title{Stringify Address Entered}
\usage{
stringify_address_entered(address_entered)
}
\arguments{
\item{address_entered}{The address entered by the user}
}
\value{
The address entered by the user formatted as a string
}
\description{
This function prints the address entered by the user.
}
