% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_module_address_entry.R
\name{stringify_address_validated}
\alias{stringify_address_validated}
\title{Stringify Address Validated}
\usage{
stringify_address_validated(address_validated)
}
\arguments{
\item{address_validated}{The address validated by the API}
}
\value{
The address validated by the API formatted as a string
}
\description{
This function prints the address validated by the API.
}
