% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mturk_assignment.R
\name{update_assignment_record}
\alias{update_assignment_record}
\title{Update Assignment Record}
\usage{
update_assignment_record(db, assignment, status, answer = NULL, attempt = F)
}
\arguments{
\item{db}{A database connection pool}

\item{assignment}{The ID of the Assignment to create a db record for}

\item{status}{The status to record for the Assignment}

\item{answer}{The Assignment answer to record. Default is \code{NULL}}

\item{attempt}{A logical indicating whether this update is the result of a review attempt}
}
\description{
Update Assignment Record
}
