% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mturk_hit.R
\name{update_hit_record}
\alias{update_hit_record}
\title{Update HIT Record}
\usage{
update_hit_record(db, hit, status)
}
\arguments{
\item{db}{A database connection}

\item{hit}{A HIT id; a string}

\item{status}{A HIT Status; a string}
}
\description{
Update HIT Record
}
