% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{update_queue}
\alias{update_queue}
\title{Update Queue}
\usage{
update_queue(db)
}
\arguments{
\item{db}{A database connection}
}
\description{
Finds and adds new cases to queue based on a case's lack of an address and presence of one or more documents
}
