% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{upload_gcs_document}
\alias{upload_gcs_document}
\title{Upload GCS Document}
\usage{
upload_gcs_document(file, id)
}
\arguments{
\item{file}{A pdf document as a raw vector}

\item{id}{The id/name of the document}
}
\description{
Upload a document to Google Cloud Storage
}
