% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgrid.R
\name{validate_postgrid_response}
\alias{validate_postgrid_response}
\title{Validate PostGrid Response}
\usage{
validate_postgrid_response(res)
}
\arguments{
\item{res}{A parsed response from the Postgrid API}
}
\value{
A list with values specified according to the PostGrid documentation
}
\description{
Validate PostGrid Response
}
