% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_session_to_db.R
\name{add_session_to_db}
\alias{add_session_to_db}
\title{Add Session to DB}
\usage{
add_session_to_db(db)
}
\arguments{
\item{db}{A database connection Pool}
}
\value{
Returns invisibly if successful
}
\description{
Adds a session to the database for {shinyauth}
}
