% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mturk.R
\name{auth_mturk}
\alias{auth_mturk}
\title{Auth MTurk}
\usage{
auth_mturk(config = NULL)
}
\arguments{
\item{config}{A file path to a config.yml}
}
\value{
A logical/boolean representing whether we successfully connected to MTurk
}
\description{
Auth MTurk
}
\examples{

auth_mturk()
auth_mturk(config = "config.yml")

}
