% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mturk_assignment.R
\name{check_assignment_record_exists}
\alias{check_assignment_record_exists}
\title{Check Assignment Record Exists}
\usage{
check_assignment_record_exists(db, assignment, hit)
}
\arguments{
\item{db}{A database connection pool}

\item{assignment}{The Assignment ID}

\item{hit}{The HIT ID}
}
\description{
Check Assignment Record Exists
}
