% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mturk_assignment.R
\name{compare_hit_assignments}
\alias{compare_hit_assignments}
\title{Compare HIT Assignments}
\usage{
compare_hit_assignments(hit)
}
\arguments{
\item{hit}{The HIT id for which to compare all assignments}
}
\value{
Nothing
}
\description{
Compare HIT Assignments
}
\examples{

\dontrun{
compare_hit_assignments(hit = "<insert hit id>")
}
}
