% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_server.R
\name{dashboard_server}
\alias{dashboard_server}
\title{Eviction Addresses Dashboard Server}
\usage{
dashboard_server(config)
}
\arguments{
\item{config}{The path to a config file}
}
\value{
A Shiny server function
}
\description{
Generates the server function passed to \code{shinyApp}
}
