% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mturk.R
\name{dispose_hit}
\alias{dispose_hit}
\title{Dispose HIT}
\usage{
dispose_hit(hit = NULL)
}
\arguments{
\item{hit}{The HIT id to dispose}
}
\value{
Nothing
}
\description{
Dispose HIT
}
\examples{

dispose_hit(hit = "<insert hit id>")

}
