% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_pay_period}
\alias{get_pay_period}
\title{Get Pay Period}
\usage{
get_pay_period(date, pay_period_start_date = lubridate::ymd("2023-01-02"))
}
\arguments{
\item{date}{A date}
}
\value{
A list with two values \code{start} and \code{end} which
are the first and last days of the pay period containing \code{date}
}
\description{
Gets the pay period for a given date
}
