% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{get_queue_length}
\alias{get_queue_length}
\title{Get Queue Length}
\usage{
get_queue_length(db, status = "available")
}
\arguments{
\item{db}{A database connection pool}
}
\value{
The length of the queue. An integer.
}
\description{
Get Queue Length
}
