% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_auth.R
\name{get_sessions_from_db}
\alias{get_sessions_from_db}
\title{Get Sessions from DB}
\usage{
get_sessions_from_db(db, cookie_expiry = 7)
}
\arguments{
\item{db}{The database connection}

\item{cookie_expiry}{The cookie expiration}
}
\value{
A tibble of session info
}
\description{
Gets a tibble of sessions from the database for use by {shinyauth}
}
