% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_users_from_db.R
\name{get_users_from_db}
\alias{get_users_from_db}
\title{Get Users from DB}
\usage{
get_users_from_db(db)
}
\arguments{
\item{db}{A database connection pool created with \code{pool::dbPool}}
}
\value{
A tibble of user info
}
\description{
Gets a tibble of users from the database to be used by {shinyauth}
}
