% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_address_validate.R
\name{handle_address_validate}
\alias{handle_address_validate}
\title{Handle Address Validate}
\usage{
handle_address_validate(db, config)
}
\arguments{
\item{db}{A database connection pool created by \code{pool::dbPool()}}

\item{config}{The path to a \code{config.yml} file to be ingested by \code{{config}}}
}
\value{
A handler function for route /address/validate
}
\description{
Handle Address Validate
}
