% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_pings.R
\name{handle_dbping}
\alias{handle_dbping}
\title{Handle DB Ping}
\usage{
handle_dbping(db)
}
\arguments{
\item{db}{A database connection pool created with \code{pool::dbPool}}
}
\value{
A plumber handler that returns a 200 status code and a message that the database is connected
}
\description{
A plumber handler that pings the database
}
\details{
This endpoint returns a simple "db pong" message after pinging the database
}
