% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_pings.R
\name{handle_dbpingfuture}
\alias{handle_dbpingfuture}
\title{Handle Future DB Ping}
\usage{
handle_dbpingfuture()
}
\value{
Empty
}
\description{
A plumber handler that pings the database in a background process, returning before returning a response
}
