% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mturk_assignment.R
\name{new_assignment_record}
\alias{new_assignment_record}
\title{New Assignment Record}
\usage{
new_assignment_record(db, hit, assignment, worker, status)
}
\arguments{
\item{worker}{}
}
\description{
New Assignment Record
}
