% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mturk.R
\name{render_hit_layout}
\alias{render_hit_layout}
\title{Render HIT Layout}
\usage{
render_hit_layout(links, layout)
}
\arguments{
\item{links}{A character vector of length greater tha zero, containing HTML links, with no missing values}

\item{layout}{A file path to a MTurk layout as specified by MTurk documentation}
}
\value{
A string (character vector length one)
}
\description{
Render HIT Layout
}
