% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mturk.R
\name{render_hit_layout}
\alias{render_hit_layout}
\title{Render HIT Layout}
\usage{
render_hit_layout(db, case, layout = NULL)
}
\arguments{
\item{db}{A database pool created with \code{pool::dbPool}}

\item{case}{A case id used to render the layout}

\item{layout}{A file path to an XML layout}
}
\value{
A character string containing an XML layout
}
\description{
Render HIT Layout
}
