% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{render_pay_report}
\alias{render_pay_report}
\title{Render Pay Report}
\usage{
render_pay_report(
  db,
  user = "jgrob",
  start = get_pay_period(lubridate::today())$start,
  end = get_pay_period(lubridate::today())$end
)
}
\arguments{
\item{db}{A database connection pool created with \code{pool::dbPool}}

\item{user}{A string giving the user to plot}

\item{start}{The start date}

\item{end}{The end date}
}
\value{
Nothing
}
\description{
Renders a pay report for a given user
}
