% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mturk_assignment.R
\name{review_assignment}
\alias{review_assignment}
\title{Review Assignment}
\usage{
review_assignment(db, config, assignment)
}
\arguments{
\item{assignment}{The Assignment id. A string (character vector length one)}
}
\value{
If successful, True
}
\description{
Review Assignment
}
