% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgrid.R
\name{send_postgrid_request}
\alias{send_postgrid_request}
\title{Send PostGrid Request}
\usage{
send_postgrid_request(address = list(), geocode = T)
}
\arguments{
\item{address}{}
}
\value{
A PostGrid response
}
\description{
Send PostGrid Request
}
