% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mturk_assignment.R
\name{update_assignment_record}
\alias{update_assignment_record}
\title{Update Assignment Record}
\usage{
update_assignment_record(db, assignment, status, answer = NULL)
}
\arguments{
\item{assignment}{}
}
\description{
Update Assignment Record
}
